requirements.txt
discord.py==2.2.2
requests==2.22.0
